Move methods from block-executor-registry to block-executor
