+++ libs/execution/src/lib/blocks/block-executor.ts
@@ -4,14 +4,27 @@
 
 import { strict as assert } from 'assert';
 
-import { IOType, isBlockDefinition } from '@jvalue/jayvee-language-server';
+import {
+  BlockDefinition,
+  IOType,
+  isBlockDefinition,
+  isCompositeBlocktypeDefinition,
+} from '@jvalue/jayvee-language-server';
 
 import { isBlockTargetedForDebugLogging } from '../debugging/debug-configuration';
 import { DebugLogVisitor } from '../debugging/debug-log-visitor';
 // eslint-disable-next-line import/no-cycle
 import { ExecutionContext } from '../execution-context';
+import { JayveeExecExtension } from '../extension';
 import { IOTypeImplementation } from '../types/io-types/io-type-implementation';
 
+import { BlockExecutorClass } from './block-executor-class';
+// eslint-disable-next-line import/no-cycle
+import {
+  createCompositeBlockExecutor,
+  getInputType,
+  getOutputType,
+} from './composite-block-executor';
 import * as R from './execution-result';
 
 export interface BlockExecutor<
@@ -79,4 +92,41 @@ export abstract class AbstractBlockExecutor<I extends IOType, O extends IOType>
     input: IOTypeImplementation<I>,
     context: ExecutionContext,
   ): Promise<R.Result<IOTypeImplementation<O> | null>>;
+}
+
+export function createBlockExecutor(
+  block: BlockDefinition,
+  execExtension: JayveeExecExtension,
+): BlockExecutor {
+  const blockType = block.type.ref;
+  assert(blockType !== undefined);
+
+  let blockExecutor = getBlockExecutorClass(blockType.name, execExtension);
+
+  if (
+    blockExecutor === undefined &&
+    isCompositeBlocktypeDefinition(block.type.ref)
+  ) {
+    blockExecutor = createCompositeBlockExecutor(
+      getInputType(block.type.ref),
+      getOutputType(block.type.ref),
+      block,
+    );
+  }
+
+  assert(
+    blockExecutor !== undefined,
+    `No executor was registered for block type ${blockType.name}`,
+  );
+
+  return new blockExecutor();
+}
+
+export function getBlockExecutorClass(
+  blockTypeName: string,
+  execExtension: JayveeExecExtension,
+): BlockExecutorClass | undefined {
+  return execExtension
+    .getBlockExecutors()
+    .find((x: BlockExecutorClass) => x.type === blockTypeName);
 }
