+++ libs/execution/src/lib/blocks/composite-block-executor.ts
@@ -20,10 +20,10 @@ import {
   isCompositeBlocktypeDefinition,
 } from '@jvalue/jayvee-language-server';
 
+// eslint-disable-next-line import/no-cycle
 import { ExecutionContext } from '../execution-context';
 import { IOTypeImplementation } from '../types';
 
-// eslint-disable-next-line import/no-cycle
 import { executeBlocks } from './block-execution-util';
 import { AbstractBlockExecutor, BlockExecutor } from './block-executor';
 import { BlockExecutorClass } from './block-executor-class';
