+++ libs/execution/src/lib/blocks/index.ts
@@ -4,6 +4,5 @@
 
 export * from './block-executor';
 export * from './block-executor-class';
-export * from './block-executor-registry';
 export * from './execution-result';
 export * from './block-execution-util';
