+++ libs/execution/src/lib/types/index.ts
@@ -3,4 +3,5 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
 export * from './io-types';
+// eslint-disable-next-line import/no-cycle
 export * from './valuetypes';
